requirements.txt
awesome-slugify==1.6.5
Django==1.10
django-braces==1.9.0
django-guardian==1.4.6
django-localflavor==1.3
django-mptt==0.8.6
django-multiselectfield==0.1.4
django-registration==2.1.2
jsonfield==1.0.3
Pillow==3.3.1
psycopg2==2.6.2
PyMySQL==0.7.7
regex==2016.9.22
requests==2.11.1
six==1.10.0
Unidecode==0.4.19
whitenoise==3.2.1
